chore(clerk-js): Move useSetSessionWithTimeout to hooks
